requirements.txt
fastapi
uvicorn
onnxruntime
opencv-python
numpy
pillow
streamlit
requests
pandas
python-multipart
